refactor(api): import getCacheControl from the env-free policy module
The v1 routes pulled `getCacheControl` through `@/config/server.mjs`, which
eagerly runs `parseServerConfig()` at module load and throws when the required
env vars are absent — so importing a route in a unit test exploded before any
test ran. server.mjs only re-exports the helper from the env-free
`@/config/policy.mjs`; import it from there instead. stats/route.js still needs
server.mjs for `config.bucketSize`.

Drops the now-dead server.mjs stub from the map route test, which can now
assert the real Cache-Control value.
