+++ src/app/api/v1/h1/stats/route.js
@@ -10,7 +10,8 @@ import { computeEtag, notModified } from '@/shared/utils/api/etag.mjs';
 import { getClientIp } from '@/shared/utils/api/clientIp.mjs';
 import { enforceRateLimit } from '@/shared/utils/api/rateLimit.mjs';
 import { backfillAndRetry } from '@/shared/utils/api/backfillSeason.mjs';
-import { config, getCacheControl } from '@/config/server.mjs';
+import { config } from '@/config/server.mjs';
+import { getCacheControl } from '@/config/policy.mjs';
 import { getStats } from '@/db/queries/getStats.mjs';
 import { umamiTrackEvent } from '@/shared/utils/umami.mjs';
 import { parseStatsQuery, projectStats } from './statsProjection.mjs';
